Use Rayon's `reduce` and im's `union_with`
This creates a new HashMap in each `tally_words` call. We can thus omit
using a Mutex. Instead of doing `for_each` with side effects, we use
`reduce` (Rayon's parallel version of it) to build the final map of
words.

The reason I used the im crate is because it has a very nice
`union_with` method on its HashMap; you can also replace `im::HashMap`
with `im::OrdMap` if you want to get sorted output directly.

I have no idea if this is faster, but it does show a different way of
using concurrency.